Take the step outline back out of the workflow editor
Opening at 100% and coming back to where a workflow was left are what
long workflows needed; the outline beside the canvas, its toggle and the
arrow keys that walked it were not.
